+++ src/nim_esphome_satellite.nim
@@ -24,6 +24,7 @@ include audio_dsp
 include xvf3800_hardware
 include wake_partition_loader
 include pcm_sound_player
+include cancellation_matcher
 
 type
   SatelliteContext* = object